refactor: section-label update paths
