+++ src/features/home/merch/merch.tsx
@@ -1,7 +1,8 @@
-import { Button, Paragraph, SectionLabel, Subtitle, Title } from '@/app/components';
+import { Button, Paragraph, Subtitle, Title } from '@/app/components';
 
 import image from '@/assets/merch.webp';
 import Image from '@/shared/image';
+import { SectionLabel } from '@/shared/section-label';
 
 import './merch.scss';
